create common utils
creating common utilities & refactoring main code into modules
