+++ src/data-quality/generic_dq_checks.py
@@ -1,4 +1,5 @@
 import json, ast, decimal
+from utils.comUtils import *
 from io import StringIO
 
 import boto3
@@ -23,43 +24,6 @@ def get_spark():
     .config("spark.jars.excludes", pydeequ.f2j_maven_coord)
     .getOrCreate())
   return spark
-
-def create_spark_df(
-  spark, source_file_path, asset_file_type,
-  asset_file_delim, asset_file_header
-):
-  if asset_file_type == 'csv' and asset_file_header == True:
-    source_df = spark.read.csv(
-      path=source_file_path, sep = asset_file_delim,
-      header=True,  inferSchema=True
-    )
-  elif asset_file_type == 'csv' and asset_file_header == False:
-    source_df = spark.read.csv(
-      path=source_file_path, 
-      sep = asset_file_delim
-    )
-  elif asset_file_type == "parquet":
-    source_df = spark.read.parquet(source_file_path)
-  elif asset_file_type == "json":
-    source_df = spark.read.json(source_file_path)
-  elif asset_file_type == "orc":
-    source_df = spark.read.orc(source_file_path)
-  try:
-    assert source_df != None
-    return source_df
-  except AssertionError:
-    print("Unable to read the data from the source")
-    
-# Helper class to convert a DynamoDB item to JSON.
-class DecimalEncoder(json.JSONEncoder):
-  def default(self, o):
-    if isinstance(o, decimal.Decimal):
-      return str(o)
-    if isinstance(o, set):  #<---resolving sets as lists
-      return list(o)
-    return super(DecimalEncoder, self).default(o)
-
-# utility function that returns spark object
 
 def main():
   spark = get_spark()
@@ -87,13 +51,15 @@ def main():
   asset_info_items = asset_info.query(
     KeyConditionExpression=Key('asset_id').eq(int(asset_id))
   )
-  print("Checkpoint!!!!!!")
-  for i in asset_info_items['Items']:
-    items = json.dumps(i, cls=DecimalEncoder)
+  items = dynamodbJsonToDict(asset_info_items)
     
-  asset_file_type = json.loads(items)['file_type']
-  asset_file_delim = json.loads(items)['file_delim']
-  asset_file_header = json.loads(items)['file_header']
+  asset_file_type = items['file_type']
+  asset_file_delim = items['file_delim']
+  asset_file_header = items['file_header']
+  
+  #asset_file_type = json.loads(items)['file_type']
+  #asset_file_delim = json.loads(items)['file_delim']
+  #asset_file_header = json.loads(items)['file_header']
   print(asset_file_type, "|", asset_file_delim, "|", asset_file_header)
 
   '''
